:truck: Move modals in their own package
