+++ src/openforms/js/components/admin/SessionStatus.js
@@ -12,10 +12,10 @@ import useTimeout from 'react-use/esm/useTimeout';
 import useTimeoutFn from 'react-use/esm/useTimeoutFn';
 import {useGlobalState} from 'state-pool';
 
+import {FormModal} from 'components/admin/modals';
 import {apiCall} from 'utils/fetch';
 import {sessionExpiresAt} from 'utils/session-expiry';
 
-import FormModal from './FormModal';
 import ActionButton, {SubmitAction} from './forms/ActionButton';
 import SubmitRow from './forms/SubmitRow';
