+++ src/openforms/js/components/admin/form_design/ChangedFormDefinitionWarning.js
@@ -3,7 +3,7 @@ import React, {useState} from 'react';
 import {FormattedMessage} from 'react-intl';
 
 import MessageList from 'components/admin/MessageList';
-import Modal from 'components/admin/Modal';
+import {Modal} from 'components/admin/modals';
 import {ChangelistColumn, ChangelistTable} from 'components/admin/tables';
 
 const AffectedFormType = PropTypes.shape({
