+++ src/openforms/js/components/admin/form_design/MissingTranslationsWarning.js
@@ -3,8 +3,8 @@ import React, {useState} from 'react';
 import {FormattedMessage} from 'react-intl';
 
 import MessageList from 'components/admin/MessageList';
-import Modal from 'components/admin/Modal';
 import TYPES from 'components/admin/form_design/types';
+import {Modal} from 'components/admin/modals';
 import {ChangelistColumn, ChangelistTable} from 'components/admin/tables';
 import jsonScriptToVar from 'utils/json-script';
