+++ src/openforms/js/components/admin/form_design/registrations/camunda/ComplexProcessVariable.js
@@ -4,7 +4,6 @@ import React from 'react';
 import {FormattedMessage, useIntl} from 'react-intl';
 import {useImmerReducer} from 'use-immer';
 
-import FormModal from 'components/admin/FormModal';
 import {SubmitAction} from 'components/admin/forms/ActionButton';
 import Field from 'components/admin/forms/Field';
 import Fieldset from 'components/admin/forms/Fieldset';
@@ -14,6 +13,7 @@ import Select from 'components/admin/forms/Select';
 import SubmitRow from 'components/admin/forms/SubmitRow';
 import {ComplexVariable, EditPanel, TypeSelector} from 'components/admin/json_editor';
 import {jsonComplex as COMPLEX_JSON_TYPES} from 'components/admin/json_editor/types';
+import {FormModal} from 'components/admin/modals';
 
 const initialState = {
   // variable definition/state
