+++ src/openforms/js/components/admin/form_design/registrations/camunda/ComplexProcessVariables.js
@@ -4,10 +4,10 @@ import {FormattedMessage, useIntl} from 'react-intl';
 import {useImmerReducer} from 'use-immer';
 
 import DeleteIcon from 'components/admin/DeleteIcon';
-import FormModal from 'components/admin/FormModal';
 import ButtonContainer from 'components/admin/forms/ButtonContainer';
 import {Checkbox, TextInput} from 'components/admin/forms/Inputs';
 import {jsonComplex as COMPLEX_JSON_TYPES} from 'components/admin/json_editor/types';
+import {FormModal} from 'components/admin/modals';
 import {ChangelistTableWrapper, HeadColumn, TableRow} from 'components/admin/tables';
 
 import ComplexProcessVariable from './ComplexProcessVariable';
