+++ src/openforms/js/components/admin/form_design/registrations/camunda/FormFields.js
@@ -3,12 +3,12 @@ import PropTypes from 'prop-types';
 import React, {useState} from 'react';
 import {FormattedMessage, useIntl} from 'react-intl';
 
-import FormModal from 'components/admin/FormModal';
 import {CustomFieldTemplate} from 'components/admin/RJSFWrapper';
 import ActionButton, {SubmitAction} from 'components/admin/forms/ActionButton';
 import Select from 'components/admin/forms/Select';
 import SubmitRow from 'components/admin/forms/SubmitRow';
 import {jsonComplex as COMPLEX_JSON_TYPES} from 'components/admin/json_editor/types';
+import {FormModal} from 'components/admin/modals';
 
 import ComplexProcessVariables from './ComplexProcessVariables';
 import SelectProcessVariables from './SelectProcessVariables';
