+++ src/openforms/js/components/admin/modals/FormModal.js
