+++ src/openforms/js/components/admin/modals/index.js
@@ -1,0 +1,2 @@
+export {default as Modal} from './Modal';
+export {default as FormModal} from './FormModal';
